Move xor helper to SDK
