+++ sdk/helper/xor/xor.go
