+++ sdk/helper/xor/xor_test.go
