+++ vault/generate_root_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/hashicorp/go-secure-stdlib/base62"
 	"github.com/hashicorp/vault/helper/namespace"
 	"github.com/hashicorp/vault/helper/pgpkeys"
-	"github.com/hashicorp/vault/helper/xor"
+	"github.com/hashicorp/vault/sdk/helper/xor"
 )
 
 func TestCore_GenerateRoot_Lifecycle(t *testing.T) {
